+++ src/database/migrations/0_0_0_010_create_activity_logs_table.php
@@ -21,9 +21,9 @@ class CreateActivityLogsTable extends Migration
             $table->json('get');
             $table->json('post');
             $table->json('files');
-            $table->bigInteger('user_id')->unsigned()->nullable();
+            $table->integer('user_id')->unsigned()->nullable();
             $table->timestamps();
-            $table->bigInteger('site_id')->unsigned()->default(1);
+            $table->integer('site_id')->unsigned()->default(1);
         });
     }
